+++ lib/wav_bytes_writer.dart
@@ -14,45 +14,69 @@
 
 import 'dart:typed_data';
 
-import 'wav_types.dart';
-import 'wav_utils.dart';
+import 'util.dart';
+import 'wav_format.dart';
 
+/// Utility class to construct a byte buffer by writing little endian ints and
+/// floats etc. Every write operation appends to the end of the buffer.
 class WavBytesWriter {
-  WavBytesWriter();
+  final _bytes = BytesBuilder();
 
-  final bytes = BytesBuilder();
+  /// Writes a Uint8 to the buffer.
+  void writeUint8(int x) => _bytes.addByte(x);
 
-  BytesBuilder writeU8(int x) => bytes..addByte(x);
-  BytesBuilder writeU16(int x) => writeU8(x)..addByte(x >> 8);
-  BytesBuilder writeU24(int x) => writeU16(x)..addByte(x >> 16);
-  BytesBuilder writeU32(int x) => writeU24(x)..addByte(x >> 24);
-  int clamp(int x, int y) => x < 0
-      ? 0
-      : x > y
-          ? y
-          : x;
-  int f2u(double x, int b) =>
-      clamp(((x + 1) * WavUtils.writeScale(b)).floor(), (1 << b) - 1);
-  BytesBuilder writeS8(double x) => writeU8(f2u(x, 8));
-  BytesBuilder writeS16(double x) => writeU16(WavUtils.fold(f2u(x, 16), 16));
-  BytesBuilder writeS24(double x) => writeU24(WavUtils.fold(f2u(x, 24), 24));
-  BytesBuilder writeS32(double x) => writeU32(WavUtils.fold(f2u(x, 32), 32));
-  final fbuf = ByteData(8);
-  void writeBytes(ByteData b, int n) => bytes.add(b.buffer.asUint8List(0, n));
-  void writeF32(double x) =>
-      writeBytes(fbuf..setFloat32(0, x, Endian.little), 4);
-  void writeF64(double x) =>
-      writeBytes(fbuf..setFloat64(0, x, Endian.little), 8);
-  void writeString(String str) {
-    for (int c in str.codeUnits) {
-      bytes.addByte(c);
+  /// Writes a Uint16 to the buffer.
+  void writeUint16(int x) {
+    writeUint8(x);
+    writeUint8(x >> 8);
+  }
+
+  /// Writes a Uint24 to the buffer.
+  void writeUint24(int x) {
+    writeUint16(x);
+    writeUint8(x >> 16);
+  }
+
+  /// Writes a Uint32 to the buffer.
+  void writeUint32(int x) {
+    writeUint24(x);
+    writeUint8(x >> 24);
+  }
+
+  void _writeSample8Bit(double x) => writeUint8(sampleToInt(x, 8));
+  void _writeSample16Bit(double x) => writeUint16(fold(sampleToInt(x, 16), 16));
+  void _writeSample24Bit(double x) => writeUint24(fold(sampleToInt(x, 24), 24));
+  void _writeSample32Bit(double x) => writeUint32(fold(sampleToInt(x, 32), 32));
+
+  void _writeBytes(ByteData b, int n) => _bytes.add(b.buffer.asUint8List(0, n));
+
+  static final _fbuf = ByteData(8);
+  void _writeSampleFloat32(double x) =>
+      _writeBytes(_fbuf..setFloat32(0, x, Endian.little), 4);
+  void _writeSampleFloat64(double x) =>
+      _writeBytes(_fbuf..setFloat64(0, x, Endian.little), 8);
+
+  /// Writes string [s] to the buffer. [s] must be ASCII only.
+  void writeString(String s) {
+    for (int c in s.codeUnits) {
+      _bytes.addByte(c);
     }
   }
 
-  SampleWriter getSampleWriter(WavFormat format) =>
-      [writeS8, writeS16, writeS24, writeS32, writeF32, writeF64][format.index];
+  /// Returns a closure that reads samples of the given [format] from this
+  /// buffer. Calling these closures advances the read head of this buffer.
+  SampleWriter getSampleWriter(WavFormat format) => [
+        _writeSample8Bit,
+        _writeSample16Bit,
+        _writeSample24Bit,
+        _writeSample32Bit,
+        _writeSampleFloat32,
+        _writeSampleFloat64,
+      ][format.index];
 
-  Uint8List takeBytes() => bytes.takeBytes();
+  /// Takes the byte buffer from [this] and clears [this].
+  Uint8List takeBytes() => _bytes.takeBytes();
 }
 
+/// Writes a sample, usually clamping it to the range [-1, 1].
 typedef SampleWriter = void Function(double);